database update(deprecated item)
git-svn-id: https://svn.omertex.com:8444/svn/iSimple@185 d8ebeebf-d64f-4e48-9a47-8608552ece91

+++ sources/isimple-project/src/com/treelev/isimple/data/BaseDAO.java
@@ -44,7 +44,7 @@ public abstract class BaseDAO {
 
     protected SQLiteStatement bindFloat(SQLiteStatement sqLiteStatement, int index, Float value) {
         if (value != null) {
-            sqLiteStatement.bindDouble(index, value);
+            sqLiteStatement.bindDouble(index, value.doubleValue());
         } else {
             sqLiteStatement.bindNull(index);
         }
